Successful implementation of temperature conversion process.
Implementation of conversion process:  Worked on lines 49-70. Essentially, depending on which symbol is more present in the file, and depending on whether it equals "C" or "F" the program will take each number within the nums array, perform the appropriate calculation on it, and shove it into the nums2[i] array, and if it doesn't equal the appropriate character, it'll set the value at i in nums2[i] equal to numbers[i], the original value in the nums array, and leave it the way it is. It'll then print each value within the nums2 array,  and the appropriate symbol that corresponds with it.